+++ src/main/java/Product.java
@@ -1,0 +1,18 @@
+package src.main.java;
+
+class Product{
+    private String name;
+    private double price;
+    public Product(String name, double price){
+        this.name=name;
+        this.price=price;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public double getPrice() {
+        return price;
+    }
+}
